requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==4.2.7
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==10.4.0
PyJWT==2.10.1
PyMySQL==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
setuptools==78.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
